# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Plataforma de gestión de asistencia, turnos, sucursales, accesos y dispositivos",
        "title": "Reloj Control API",
        "termsOfService": "https://minolsoft.cl/terms",
        "contact": {
            "name": "Minolsoft",
            "url": "https://minolsoft.cl",
            "email": "soporte@minolsoft.cl"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://minolsoft.cl/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/shifts/{id}/days": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lista la configuración semanal del turno. PUT reemplaza toda la configuración semanal del turno.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift Days"
                ],
                "summary": "Días del turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reemplazar días del turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.shiftDayInput"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ShiftDayDTO"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lista la configuración semanal del turno. PUT reemplaza toda la configuración semanal del turno.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift Days"
                ],
                "summary": "Días del turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reemplazar días del turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.shiftDayInput"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ShiftDayDTO"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
...
        "handlers.ShiftDayDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_working_day": {
                    "type": "boolean",
                    "example": true
                },
                "mode": {
                    "type": "string",
                    "example": "onsite"
                },
                "shift_id": {
                    "type": "integer",
                    "example": 1
                },
                "weekday": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.shiftDayInput": {
            "type": "object",
            "properties": {
                "is_working_day": {
                    "type": "boolean",
                    "example": true
                },
                "mode": {
                    "type": "string",
                    "example": "onsite"
                },
                "weekday": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}